fix(kornia-py): fix DLPack device-import UAF; drop copy arg (always zero-copy)
Memory-safety fix + API simplification in the device DLPack import path, plus a
first step of splitting the oversized cuda.rs.

- UAF fix: dlpack_to_device_arc borrowed the producer's __dlpack__ capsule but
  never consumed it, so on return the capsule GC'd and ran the producer's
  deleter at import time. For a spec-conformant producer that hands buffer
  ownership solely to the managed tensor, the zero-copy alias then dangled
  (use-after-free; torch/cupy were only accidentally safe because they retain an
  independent ref). Now consume the capsule (PyCapsule_SetName -> used_dltensor)
  exactly like the host Image::from_dlpack path, so the deleter can't fire early
  and the buffer stays alive via the obj keep-alive.
- Remove the `copy` argument from Image.from_dlpack (import): always zero-copy,
  mirroring torch's DLPack. Deletes the copy=true device-to-device branch; the
  imported alias is read-only. Callers wanting an owned buffer copy it
  themselves. (`from_numpy`'s copy and the `__dlpack__` protocol kwarg are
  unchanged.)
- Extract the GPU color ops into cuda_ext/cuda_color.rs (cuda.rs -> cuda_ext/mod.rs)
  as the first step toward breaking up the 1800-line file.

Verified on both feature configs (build + clippy) and the device/torch-interop/
DLPack suites (63 passed, only the known Jetson mem-probe noise-fail); torch
zero-copy import round-trip confirmed on the system torch-CUDA build.
